+++ src/packages/eslint-config-kiwicom/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Add `FormData` to the globals
+- Rule `import/no-unresolved` now ignores Relay artifacts set by `artifactDirectory`
 
 # 2.7.0
 - Enable warnings for `jest/prefer-todo` rule